implement WholeBodyControllers library
