+++ cmake/WalkingControllersFindDependencies.cmake
@@ -119,6 +119,9 @@ endmacro()
 find_package(YARP QUIET)
 checkandset_dependency(YARP)
 
+find_package(ICUB QUIET)
+checkandset_dependency(ICUB)
+
 find_package(iDynTree QUIET)
 checkandset_dependency(iDynTree)
 
@@ -128,9 +131,14 @@ checkandset_dependency(osqp)
 find_package(OsqpEigen 0.4.0 QUIET)
 checkandset_dependency(OsqpEigen)
 
+find_package(qpOASES QUIET)
+checkandset_dependency(qpOASES)
+
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libYARP_helper "Compile libYARP_helper?" ON WALKING_CONTROLLERS_HAS_YARP OFF)
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libiDynTree_helper "Compile libiDynTree_helper?" ON "WALKING_CONTROLLERS_HAS_iDynTree;WALKING_CONTROLLERS_HAS_YARP" OFF)
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libSimplified_model_controllers "Compile libSimplified_model_controllers?" ON
                                     "WALKING_CONTROLLERS_COMPILE_libYARP_helper;WALKING_CONTROLLERS_COMPILE_libiDynTree_helper;WALKING_CONTROLLERS_HAS_osqp;WALKING_CONTROLLERS_HAS_OsqpEigen" OFF)
 walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libRobot_helper "Compile libRobot_helper?" ON
                                     "WALKING_CONTROLLERS_COMPILE_libYARP_helper;WALKING_CONTROLLERS_COMPILE_libiDynTree_helper" OFF)
+walking_controllers_dependent_option(WALKING_CONTROLLERS_COMPILE_libWhole_body_controllers "Compile libWhole_body_controllers?" ON
+                                    "WALKING_CONTROLLERS_COMPILE_libYARP_helper;WALKING_CONTROLLERS_COMPILE_libiDynTree_helper;WALKING_CONTROLLERS_HAS_osqp;WALKING_CONTROLLERS_HAS_OsqpEigen;WALKING_CONTROLLERS_HAS_qpOASES;WALKING_CONTROLLERS_HAS_ICUB" OFF)
