+++ src/CMakeLists.txt
@@ -8,3 +8,4 @@ add_subdirectory(libTime_profiler)
 add_subdirectory(libStd_helper)
 add_subdirectory(libSimplified_model_controllers)
 add_subdirectory(libRobot_helper)
+add_subdirectory(libWhole_body_controllers)
